requirements.txt
requests
aiohttp
python-telegram-bot==20.7
